requirements.txt
python==3.7
pytorch==1.7.1
numpy==1.20.3
scipy==1.7.1